Change -sim option with new "all".
This separates "all" from "stage1", which now allows for starting to
resim on stage1 without re-generating the permutated profiles (which
might be costly as well).
So in practice if you specify -sim stage1 you will be asked if you want
to regenerate "out.simc", similar to how it already is on stage2/3.
For dynamic_stage1 without having regenerated profiles, this means we do
not know the number of generated profiles and thus can not offer
estimation times.

Continue fight against globals.