Allow for N stage simulations (!=3).
This is not yet fully and perfectly fleshed out for all use cases.
It is also only accessible in settings.py for more advanced users.
Can/will be extended more with better user input() and sensible defaults
to maybe even have dynamic "optimal" values of target_error, etc.
depending on number of stages and number of profiles generated.